Addressed review comments.
Moved methods to appropriate classes.
